refactoring contd..
