+++ internal/aws/filter_test.go
@@ -1,4 +1,4 @@
-package aws_new
+package aws
 
 import (
 	"github.com/aws/aws-sdk-go-v2/aws"
