+++ internal/aws/get_cost_and_usage.go
@@ -1,4 +1,4 @@
-package aws_new
+package aws
 
 import (
 	"context"
