+++ internal/cli/flags/dimension_and_tag_filterby.go
@@ -19,11 +19,11 @@ func NewFilterBy() DimensionAndTagFilterFlag {
 
 func (f *DimensionAndTagFilterFlag) Set(value string) error {
 
-	args := utils_new.SplitCommaSeparatedString(value)
+	args := utils.SplitCommaSeparatedString(value)
 
 	for _, arg := range args {
 
-		parts, err := utils_new.SplitNameValuePair(arg)
+		parts, err := utils.SplitNameValuePair(arg)
 		if err != nil {
 			return err
 		}
