+++ internal/cli/flags/dimension_and_tag_groupby.go
@@ -16,11 +16,11 @@ func (e DimensionAndTagFlagError) Error() string {
 
 func (f *DimensionAndTagFlag) Set(value string) error {
 
-	args := utils_new.SplitCommaSeparatedString(value)
+	args := utils.SplitCommaSeparatedString(value)
 
 	for _, arg := range args {
 
-		parts, err := utils_new.SplitNameValuePair(arg)
+		parts, err := utils.SplitNameValuePair(arg)
 		if err != nil {
 			return err
 		}
